+++ boot.php
@@ -1,0 +1,6 @@
+<?php
+
+date_default_timezone_set("Europe/Moscow");
+
+require('./functions/helpers.php');
+require('./functions/main.php');
